+++ src/test/java/io/codemine/java/postgresql/codecs/OidCodecTest.java
@@ -2,7 +2,7 @@ package io.codemine.java.postgresql.codecs;
 
 import io.codemine.java.postgresql.CodecTestBase;
 
-public class OidCodecTest extends CodecTestBase<Integer> {
+public class OidCodecTest extends CodecTestBase<Long> {
   public OidCodecTest() {
     super(Codec.OID);
   }